patch to allow larger stdfs, tons of comments added to the v4 records, added vur/psr types; added a "die grouping" method that groups stuff up for you; added a cache for non die level records; added a PMR lookup class and an extension on mprs that uses it; added round trip tests for the new record types; added the name of the variables in the spec to the metadata for the classes; test program just has some random stuff that i used to make sure the stuff i added works. the stdf file summary creator was the first thing i did with the linqtostdf package so dont judge me too hard for it. including it as an example utility.

+++ Main/LinqToStdf/RewindableByteStream.cs
@@ -30,7 +30,7 @@ namespace LinqToStdf {
             /// <summary>
             /// Tracks our current offset in the stream
             /// </summary>
-            int _Offset = 0;
+            long _Offset = 0;
 
             /// <summary>
             /// buffer used for reading a record header
@@ -189,9 +189,9 @@ namespace LinqToStdf {
             }
 
             /// <summary>
-            /// The current offset
+            /// The current offset (switched to long so it doesn't limit the size of stdf files being inputted)
             /// </summary>
-            public int Offset {
+            public long Offset {
                 get {
                     return _Offset;
                 }
